requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.0.0
qrcode==8.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
drf-spectacular==0.28.0
# Uncomment below for PostgreSQL:
# psycopg2-binary==2.9.10
# Uncomment below for Celery/Redis background tasks:
# celery==5.4.0
# redis==5.2.1
